Baseline_code/Baseline/code/counterfactual_generative_networks-main/condaenv.xww99sxo.requirements.txt
boto3==1.16.56
botocore==1.19.56
chardet==4.0.0
cycler==0.10.0
defusedxml==0.6.0
dill==0.3.3
entrypoints==0.3
filelock==3.0.12
gdown==3.12.2
idna==2.10
imageio==2.9.0
ipdb==0.13.4
ipykernel==5.4.3
jmespath==0.10.0
jupyter-client==6.1.11
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
notebook==6.2.0
numpy==1.19.5
pandas==1.2.1
parso==0.8.1
prompt-toolkit==3.0.10
protobuf==3.14.0
pysocks==1.7.1
pytz==2020.5
pywavelets==1.1.1
pyyaml==5.3.1
pyzmq==21.0.1
qtconsole==5.0.1
repackage==0.7.3
requests==2.25.1
s3transfer==0.3.4
scikit-image==0.18.1
scipy==1.6.0
tensorboardx==2.1
terminado==0.9.2
tifffile==2021.1.14
torchsummary==1.5.1
tqdm==4.56.0
urllib3==1.26.2
webencodings==0.5.1
yacs==0.1.8